bootstat.cpp: remove record for time_since_factory_reset
time_since_factory_reset is set with the difference between
current time(boot complete time) and the factory reset time
which is only recorded at the first boot.
There is no meaning to compare the boot complete time with
boot time for the first boot after factory reset for every
boot, so remove it with this change to avoid confusion.

Test: manually with HiKey

Change-Id: Ibb0bf561b438be72220506aedbab3e6008b21c07
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>