Update patch set 1
Patch Set 1:

> > the current name seems much clearer to me.
 > 
 > hmm, this is the output I got from the second boot with hikey
 > device:
 > hikey:/ # bootstat -p
 > Boot events:
 > ------------
 > last_boot_time_utc 32
 > build_date 1498757411
 > factory_reset_boot_complete_no_encryption 62
 > factory_reset_boot_complete 62
 > ro.boottime.init 5769
 > ro.boottime.init.selinux 154
 > ro.boottime.init.cold_boot_wait 330
 > boottime.bootloader.total 0
 > absolute_boot_time 26
 > boot_reason 1
 > factory_reset_current_time 32
 > factory_reset 68
 > time_since_last_boot -36
 > boot_complete_no_encryption 26
 > boot_complete 26
 > factory_reset_record_value 68
 > time_since_factory_reset -36
 > hikey:/ #
 > 
 > time_since_last_boot is a negative value there, which seems not as
 > expected.

It's a known issue that the clock may not be set by the time this value is calculated.  Potential solutions for that are adding a signal from the platform that the clock is set or adding a time delay to try to increase the chance that the clock is set when this metric is calculated (though that's hacky, not favorable).

Patch-set: 1
Reviewer: Gerrit User 1081159 <1081159@85c56323-6fa9-3386-8a01-6480fb634889>