Update patch set 2
Patch Set 2:

(4 comments)

NDK is supposed to be specified in liblog.map.txt AFAIK, which generates the stub liblog.so that NDK products can be compiled against, and does not list this symbol. The symbols are not referenced in the only NDK include android/log.h.

It has been available for a long time, for _internal_ in-tree use.

I smell a lot of issue with technical debt ...

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>