Update patch set 1
Patch Set 1:

Talked with enh@ in person a bit.  Thoughts are to keep with milliseconds.  The extra fidelity isn't likely to help anyone and there are already overflow problems with std::chrono, so no need to compound them with a finer grained type.

We'll keep operator<< as is right now and replace with PrettyDuration in the future, as well as add operator<< for the major std::chrono types.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>