Update patch set 1
Patch Set 1:

My sailfish looks like it's printing things endlessly when it hits this code (before the CL that is), but I don't see a loop here.  I know that dumping all of these stacks is going to take a while, but are you seeing something similar.

This is particularly in the disable verity / remount case, before your other CL was merged.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>