Update patch set 1
Patch Set 1:

This came up in a code review, so uploaded here.

I'm not sure that I like it though.  The real answer here is that there should be a ServiceBuilder and all of the Service::Parse* functions should be factored out into a ServiceParser...  Is that overkill? Maybe, but it has a use too.  For host parsing of init scripts, I explicitly do not need any of the functionality provided by the Service class; I only need the parsing and breaking this up appropriately will allow a host init parser without bringing in the Service dependencies (some of which are problematic).

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>