[WIP] Add an alternative DT dir for early mount
On platforms that use ACPI instead of Device Tree (DT), such as
Ranchu x86/x86_64, /proc/device-tree/firmware/android/ does not
exist. As a result, Android O is unable to mount /system, etc.
at the first stage of init:

 init: First stage mount skipped (missing/incompatible fstab in
 device tree)

Instead of printing that warning and giving up, make /init look in
another directory, which is based in sysfs and laid out like the
standard DT directory in procfs, for early mount configuration. This
alternative DT directory is created by the goldfish_fwdata driver
using information in ACPI tables:

 https://android-review.googlesource.com/442472
 https://android-review.googlesource.com/442393
 https://android-review.googlesource.com/442395

With this CL as well as the above ones, the said warning is gone,
but early mount fails, and Android Emulator keeps rebooting:

 init: bool android::init::FirstStageMount::InitRequiredDevices():
 partition(s) not found in /sys, waiting for their uevent(s): vda
 init: Wait for partitions returned after 10010ms
 init: bool android::init::FirstStageMount::InitRequiredDevices():
 partition(s) not found after polling timeout: vda
 init: Failed to mount required partitions early ...
 init: panic: rebooting to bootloader

Test: Boot patched sdk_phone_x86-userdebug system image with patched
      Android 3.18 x86 kernel in patched Android Emulator, verify
      the "init: First stage mount skipped" warning no longer shows
      in dmesg.

Change-Id: Ib6df577319503ec1ca778de2b5458cc72ce07415
Signed-off-by: Yu Ning <yu.ning@intel.com>