Update patch set 1
Patch Set 1:

Essentially you're saying that the emulator is failing to boot because it's failing to find partitions to mount, right?

There is code in init that gets the partitions to mount, GetRequiredDevices() in init_first_stage.cpp.  This should return no partitions if there are no partitions ready to be mounted.  That is a better fix here.

Otherwise, when are the partitions being formatted?  Can they be formatted before the boot is attempted?

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>