Update patch set 1
Patch Set 1:

> Essentially you're saying that the emulator is failing to boot
 > because it's failing to find partitions to mount, right?
 > 
 > There is code in init that gets the partitions to mount,
 > GetRequiredDevices() in init_first_stage.cpp.  This should return
 > no partitions if there are no partitions ready to be mounted.  That
 > is a better fix here.
 > 
 > Otherwise, when are the partitions being formatted?  Can they be
 > formatted before the boot is attempted?

the system.img is read only and it will not be formatted once created.
since by creation, there is no partition on it, it just stay that way.

in api25 and before, this is not a problem; even in O where there is
no treble involved (no first stage mount), this is also not a problem:
init just happily mount it without problem. IIUC,   mount(...) system
call is working fine.

it appears to be some assumption made in
first stage mounting that there has to be some partitions on the disk.

I am not sure how to remove that assumption in first stage mount.

Patch-set: 1