Update patch set 4
Patch Set 4:

Yes, this latest patch is what I was hoping for.  It solves a lot of the problems of the previous ones.  

The only comment I have is that we should use std::call_once() for the lazy initialization in case any users of these functions end up multithreaded.  Looks good to me otherwise.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>