Update patch set 2
Patch Set 2:

I was trying to do this in a way that didn't need Service::Start() to need to reference ServiceManager, since services get started independently of ServceManager.  I didn't want to store a pointer to the parent and I didn't want to abuse the singleton anymore, as I'm totally going to get rid of it (some day... really).

I figure the filter and sort only in shutdown is cheap enough.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>