logging: Make clion friendly
* Force CLion to see DCHECK as always being executed,
  by making sure that it ignores NDEBUG.

* Update definition of CHECK(x) and friends to make them
  friendly to clion's code inspector.

* Update definition of LOG(x) and friends.

Previously it used to think "x || cout << x" had an unused
expression value (which is how CHECK, but not CHECK_OP expands).

Since CLion ignores _Pragma push/pop, we cannot push a temporary
"-Wunused-value" pragma.

Instead, redefine CHECK(x) as "if(!x)..." instead of "x || ..."

LOG(x) is similarly redefined as "if(x)..." instead of "x && ..."

(As a result, the # of clion warnings for ART code drops by about 90%).

Test: Manually looking at ART source code in CLion
Bug: b/64455231
Change-Id: I3699b7f77d297f40fc54845575998acae321f5be