Update patch set 1
Patch Set 1:

I've been meaning to do something like this for a while actually.  I hit some linker errors where sehandle wasn't in libinit and that finally broke me to doing it.

It's not quite done yet, but it does compile.  I need to clean up a lot of the headers, but I'm looking for reviews in the meantime particularly on the little documentation blurb I wrote at the top of selinux.cpp.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>