Update patch set 1
Patch Set 1:

(4 comments)

Thanks for review to both of you.

I decided that bootstat is shipped always, and _can_ be made part of the critical path, at least in this prototype to explore ways and means of refining the boot reason to a standardized, minimal and meaningful set. This is the _only_ application or daemon that already had built-in sensitivity to ota and factory_reset; two boot reasons missing from all other APIs. Two phases of the several gave the answers to all.

I would like to know if bootstat will remain a permanent fixture though. (comment in document may be better than in AOSP).

I do not believe in creating more daemons, when existing ones can be used/enhanced without much harm (remember, this is still a prototype). I also do not believe in proliferating APIs, which has happened with boot reasons (sys.boot.reason is Yet Another API, that does not escape irony).

The next step may be to explore controlling the content in /data/misc/reboot/ to make that the API.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0