DO NOT SUBMIT: Create new BacktraceMap iterator.
This allows the new unwinder code to avoid having to create a duplicate copy
of the map data. Only when the map is accessed is the backtrace_map_t created.

This is just an experiment and is probably not worth it. A better idea might
be to rewrite libbacktrace to be a much thinner wrapper for libunwindstack.

Bug: 23762183

Test: Unit tests pass.
Change-Id: I43d81429d72c7096e7056eb77d62489826b1556b