Update patch set 2
Patch Set 2:

This isn't anywhere near ready to submit, but there's somewhat of a mismatch between the libbacktrace API and what I want to do.

I have a bunch of thread ucontexts, and a process that's a fork of the address space corresponding to the threads. libbacktrace doesn't seem to provide a way to actually do this, other than maybe lying to it and giving the vm fork pid as the pid. Is this something that you think is a reasonable thing to do?

Patch-set: 2
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>