Update patch set 2
Patch Set 2:

(9 comments)

Thanks!

Glad you also covered the logdOpen part too, I did not think about it in my request to fix this, so good catch.

I have a lot of KISS suggestions, rationalizing the removal of the remaining legacy of EINTR special checking, which does not make sense w/o alarm() in play here.

However, testing concern, we should _manually_ check if SO_SNDTIMEDOUT and poll works if logd crash tombstones are to be collected. The issue is a deadlock loop because part of the content in the tombstones is the associated logcat (which can not happen from logd, none is generated, none can be collected). The issue is that once the process is paused and being PTRACED to grab the stack trace and other process details before allowing it to be removed from the process list, the writer and reader ends of the UNIX domain socket may not work as 'expected' (their behavior may be paused too, easier to test than to inspect the kernel code).

Only userdebug and eng builds allow logd to be PTRACED because of security concerns, so this is _purely_ a debugging issue. No crash reports for logd on user builds ...

Or maybe, it is time to consider a gTest that checks this destructive tombstone/crash scenario?

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>