logd: remove logd-reinit
Waiting for ro.persistent_properties.ready to be true and
persist.logd.size to change achieves the same with less complexity.

More importantly, it allows us to simply check this property at the
startup of logd, which allows logd to conditionally load from /data
based on whether or not it has been mounted, and therefore fixes a
race condition that can cause an SELinux denial during boot.

Bug: 64723606
Test: boot bullhead, logd unit tests, liblog unit tests
Test: check that changing the log buffer size in settings changes the
      value seen in `logcat -g`
Change-Id: I8335b0af25d347c997b1780f0693c627f75ec1be