Update patch set 2
Patch Set 2:

(1 comment)

> The reinit command could take an argument to be more precise, and Setup.cpp could just do something like a write to "/dev/socket/logd" with the string "reinit size\0EXIT\0" instead of communicating to init to trigger the logd-reinit command. with a 'size' argument, only the sizes are re-initialized?

Are you saying this regarding the settings layer?

I see that it sets a few properties but it always sets the SELECT_LOGD_SIZE_PROPERTY property.  Perhaps instead of the reinit logic, we could simply have a thread waiting for a change to this property, and whenever that property gets changed, we do the logBuf->init() logic?

In theory, it'd also allow shell users to change SELECT_LOGD_SIZE_PROPERTY and trigger the same reinitialization in logd automatically.

Is there anything I'm missing for why this might not work?

Patch-set: 2