Update patch set 3
Patch Set 3:

> Automatic reinitialization for property changes is great! Except there are _many_ more properties to watch/view (logd/README.properties) that are associated with buffer size if one is to support 'random' shell adjustments. 

I'm thinking of just triggering off of SELECT_LOGD_SIZE_PROPERTY aka persist.logd.size, since that is the one that is unconditionally set before doing logd --reinit.  It'd keep the current Settings behavior intact.

What happens currently if the other of these properties are set?  One needs to reboot (for the persistent properties) or otherwise call logd --reinit manually?

> Is the expense of watching a property low, I have not inspected the mechanism. 

It's just a futex_wait(), so waiting for a single property is very efficient.

> If it spawns a thread(s) to do so, we would like them prctl'd as it has been immensely useful to pinpoint logd CPU usage blame.

I'd use the same thread that I'm replacing here.

The only caveat, is that I can't wait on multiple properties easily.  I plan in the future to add that mechanism to property service, but that's not ready yet.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>