Update patch set 1
Patch Set 1:

I worry a bit about borrowing the socket for SIGCHLD to also handle SIGTERM as there isn't a great recovery mechanism if the read() fails and we get partial data (even if that's unlikely).

This can likely be done cleanly with signalfd().  The new fd can be added to the epoll set and when it fires, a read() will return signalfd_siginfo that contains ssi_pid for checking against 0 then rebooting.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>