Update patch set 2
Patch Set 2: Presubmit-Verified+1

=== Finished presubmit run: L59600000102890162 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/485658?revision=2




TreeHugger finished with: 3 passed.



WARNING: The tip-of-tree(4339199) for aosp-master/aosp_arm64-eng may be broken.
The presubmit build passes with the last-known-good-build(4337575), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4339199) for aosp-master/aosp_x86-eng may be broken.
The presubmit build passes with the last-known-good-build(4337034), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



I'm your biggest fan!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 2
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1