init: fix shutdown deferral
In the previous change that defers shutdowns requested by builtin
functions, there was an issue: if the last command in the action queue
requests the shutdown or if that command starts a property wait or
start an exec service, epoll_timeout_ms is not set to 0, which is
needed to process the shutdown request at the beginning of the next
iteration of the loop.

This change rectifies the issue by explicitly setting epoll_timeout_ms
to 0 if do_shutdown is set.

Bug: 65374456
Test: reboot bullhead from the last command in the action queue
Change-Id: Idc79f084610f084c3dfd7761f76ac907fa8d2c44