Update patch set 2
Patch Set 2:

> > This probably isn't strictly legal, but I don't think it's likely
 > that it will break.
 > 
 > Any references to the spec I should quote here?

C++17 20.5.4.2.1 Namespace std [namespace.std]

The behavior of a C++ program is undefined if it adds declarations or definitions to namespace std or to a
namespace within namespace std unless otherwise specified. A program may add a template specialization
for any standard library template to namespace std only if the declaration depends on a user-defined type
and the specialization meets the standard library requirements for the original template and is not explicitly
prohibited.

Patch-set: 2
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>