Update patch set 4
Patch Set 4:

(1 comment)

I had to make the change to the battery code, because the ber-fuzzy rfind would work in noise, and as such greatly increase the chances that the interpretation of the battery percentage would _also_ be full of noise.

refining/deprecating battery code is outside the scope of this change. I am not ready to throw out the baby with the bathwater yet because it is itself an experiment and I still think it may evolve to something useful.

Patch-set: 4
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0