Update patch set 5
Patch Set 5:

I found only _two_ bonafide cases of existing code where this made sense so far (bootanim and dumpstate).

For the cases where reads are fulfilled incrementally and sent via WriteFully I could not conscientiously replace. For instance, that was what was going on in dumpsys (and several others). I had troubles with using this function as it would read the entire content in, swallowing up memory to do so, _then_ propagate the content. As-is the code in dumpsys would not risk introducing memory pressure right when you are collecting a bugreport ...

adb presented a difficulty because it had wrappers to support WIN32 that complicated replacing anything there. As such we will need another spin to ensure _this_ code is portable to all operating systems in any case, so we are not done yet.

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>