Update patch set 6
Patch Set 6:

> chrono literals were supported in C++11 according to the include file ;-/

Strange; the documentation all says they're added in C++14.

> Drop 0ms check, and adjust line 101:
> if ((remaining_timeout.count() <= 0) && (n != BUFSIZ)) return true;

Do you mean line 128 (PS6) ?

This is what I'd prefer as a solution here.  The strange situation is what happens with timeout = 1ms when we're reading a large file?  Do we only get the amount that can be read in 1ms worth of poll()'s succeeding?  That seems to be wrong; I'd expect that the timeout is waiting for data to be available, not also including the time of the reads, which this change would also fix.

But... if we're needing to talk about these assumptions and some of the potential users (uevent_listener.cpp, property_service.cpp) already have their own slight tweaks to this behavior, then it's making me think that this isn't a good candidate for libbase.  One of the major goals of libbase is to provide simple functionality that everyone's already using and has very few assumptions, and unfortunately, we're being forced to put many assumptions into this behavior.

Patch-set: 6
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>