init: support suspend and hibernation
Create shutdown,suspend and shutdown,hibernate request forced suspend
to RAM or DISK respectively.  Should only be used after necessary high
level preparation and cleanup have occurred.

The request to the kernel to perform the suspend or hibernate
operations are gated with "ro.support.suspend" and
"ro.support.hibernate" boolean properties respectively, the platform
enables these respective requests by setting them to "true".  If
suspend is disabled, it will do nothing.  If hibernate is disabled,
it will do a full umount and fsck shutdown so that boot up is as
short as possible.

Caller is not informed of success or failure.  Caller must make its
own arrangements to determine if the device did suspend or hibernate,
and upon return evaluate if it should retry.

FutureProofing: Besides "true" or "false", the support properties may
change from Bool to String to select alternate default mechanisms to
perform the requests.  A third field in the request may also serve
this purpose so runtime can select.  For instance in the future
"autosleep" may switch the request from immediate (and possibly fail
or be rejected, to be retried later), to deferred until all systems
and drivers have released their wake locks and reference counts.

Test: manual
Bug: 63736262
Bug: 38446744
Bug: 66912053
Change-Id: I1f3911269e869a24d700e11c03a838a18da3b3a9