init: support suspend and hibernation
Create pseudo shutdown,suspend and shutdown,hibernate requests, making
init the hardware abstraction layer to request these operations and to
perform a forced operation with any necessary low-level cleanup.  The
request to the kernel to perform the suspend or hibernate operations
are gated with "ro.support.suspend" and "ro.support.hibernate"
properties respectively, it is the responsibility of the platform to
enable these if required by setting the property to a value of 1,
otherwise they default to disabled.

Test: compile
Bug: 63736262
Bug: 38446744
Bug: 66912053
Change-Id: I1f3911269e869a24d700e11c03a838a18da3b3a9