init: shutdown,[suspend|hibernate] turn off backlight
Refactor TurnOffBacklight into SetBacklight that returns the previous
value for the backlight so that we can in return reset the value.  For
shutdown,suspend and shutdown,hibernate we will turn off the
backlight before asking kernel to forcibly perform the related
operations.  It is expected that a higher level power manager will
arrange for drawing blanked content to the screen and other cleanup,
the only reason we turn off the backlight is to act as a failsafe.

It should be noted that turning off the backlight does not mean the
content is not visible, for instance in strong lighting, so it is
an imperfect failsafe.  Better than suspending the device with
frozen content clearly visible on the screen at full brightness.

Test: manual
Bug: 63736262
Bug: 38446744
Bug: 66912053
Change-Id: Ie61c3bf6cffda27628a0eb91eb9cae05178ba42d