Update patch set 2
Patch Set 2:

I'm fine with renaming this, but I'm not sure sigterm is the right name.  This handler is to allow a host OS to ask an Android instance running in a container to shutdown, so while the mechanism is sigterm, that's not really a good indicator of the actual reason.  

Luis, who added this code, can probably advise better for the name.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>