Update patch set 2
Patch Set 2:

The way I'm reading this code, it looks like we fully intend to leak this ResolvedService object. Rather than jumping through hoops to try to confuse the analyzer about that, why can't we just add a comment that we're leaking this and put a NOLINT here?

Patch-set: 2
Reviewer: Gerrit User 1096946 <1096946@85c56323-6fa9-3386-8a01-6480fb634889>