Update patch set 4
Patch Set 4:

libsuspend is not considered, although it can, leaving that for later. I was expecting the API should only be in the PowerManager(s), effectively setprop sys.powerctl shutdown,suspend which it uses today for all other power manager operations. I assume once it has made any of the necessary arrangements to notify applications of the impending operation. I also expect that the wakeup sources to be adjusted and limited to the KeySwitch (in the case of Android Auto) and none of the other normal suspend wake events (Alarm, usb connection change, etc).

Patch-set: 4
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0