init: shutdown,[suspend|hibernate] turn off wake sources
Add heuristic to suspend all wakesources, exclude any interrupts
tied to switches and keys.  This will permit a power key to wake us
up.  We can not control wake sources that have no means of being
disabled, which is common for the modem and radio subsystem, so
caller is responsible for telling radios to turn themselves off.

Test: manual
Bug: 63736262
Bug: 38446744
Bug: 66912053
Change-Id: I0c2881a68f9255dcdb4ab8777a74e83d54d84cae