Update patch set 15
Patch Set 15: Presubmit-Ready+1

Any other comments on this?  It's ready to be submitted from my perspective.  The latest CL removes the line to actually enable the vendor subcontext, so it is effectively a no-op, but would get the mechanism in place while we work on the SEPolicy.

My sailfish is booting > 100 cycles with the subcontext enabled (albeit in permissive mode)

Patch-set: 15
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1