Update patch set 16
Patch Set 16:

> > Unless... we assume that we only reference the arguments in
 > `Command` more than once and therefore it'd be fine to just have an
 > accessor that provides access to args(x) and does the expand before
 > returning?  Let me look at doing this.
 > 
 > Nevermind, this is a bad idea from the start.  I don't recreate a
 > `Command` object after sending a message to the subcontext and we
 > proactively return a failure instead of calling the function if any
 > of its arguments cannot be expanded.  Making this change would
 > remove that proactive error.
 > 
 > I can still build a data class if that'd be prettier, but it'd
 > literally just be a data class.

i think normally it would be 3 arguments before i said "this should be an object", but `context` just seems so random it's already feeling like an object rather than things that obviously belong together.

i think the best counter-argument would be if you have -Werror=unused-parameter and are using this to force builtins to actually pay attention, but it seems like most are just silently ignoring `context` so it doesn't seem like there's a "please be careful" benefit to not just hiding it in an object.

Patch-set: 16
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>