adb: support [test|wait]-for[-TRANSPORT][-not]-STATE[-TIMEOUTms]
Allow scripting to bail out on wait for state changes.  Return an
error reason of "timeout".  A timeout of 0ms can be used as a means
to test if we are in, or not in, the associated state; to that ends
we also added test-for[-TRANSPORT][-not]-STATE as a shorthand that
makes more sense.

Test: manual
Bug: 63736262
Bug: 38446744
Bug: 66912053
Change-Id: Idbde65a75781daaa0b82477d7785f287ec568528