Update patch set 2
Patch Set 2:

Upon inspection of the documentation, there appears to be reference to an adb connection to the bootloader. Indicating it is an unique state where only one communications stream is allowed and it is in an "embedded" state. Is this for realz?

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>