Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> 1) will we support executables in /odm/bin? 
> - if we want to have aditional services stored in odm we have to add the execustables in the path
> 

ok

> 2) will we support loading libs from /odm/lib in processes started from /system/bin? 
> - the executables stored on odm have aditional dependencies that are stored in /odm/lib
> 

This doesn't answer to my question. I am asking whether there will be a dependency from the system partition to the odm partition or not.

> 3) will we support loading libs from /odm/lib in processes started from /vendor/bin? 
> - most probably yes since some extensions will be stored on /odm
> 

This doesn't seems to be correct. You are saying that vendor -> odm dependency will exist and this implies that vendor has to have a knowledge on odm which is a downstream of vendor.

> 4) will we support loading libs from /vendor/lib in processes started from /odm/bin? 
> - due to soc support this is feature needed
> 

Ok. But this says that your odm partition is tied to a specific vendor; Many of the device manufacturers are using SoC chipsets from different vendors for a single family of devices depending on the market. In that case, the odm -> vendor dependency will be a problem. 

> 5) what will be the priority in between /system and /odm partitions? If /odm/lib comes prior to /system/lib, we can use /odm/lib to override some libs in /system/lib. However, this also looks dangerous unless /odm partition is protected as securely as /system partition (i.e. /odm also needs to be dm-verified)
> - in the test /odm was set with lowest priority and it could not override the libs or services from /system or /vendor

Ok.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>