Update patch set 4
Patch Set 4:

This will allow test scripting to use adb to perform the timeout rather than dedicating a thread to the operation and another to managing a timeout to deal with erroneous device behavior.

adb wait-for-device-30000ms can be used to continue and propagate an error rather than sitting forever waiting for something the device could be in a state where it will never provide.

Patch-set: 4
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>